Ticks W.I.P.

+++ includes/vector.h
@@ -6,7 +6,7 @@
 /*   By: chamada <chamada@student.le-101.fr>        +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/12/02 13:22:33 by chamada      #+#   ##    ##    #+#       */
-/*   Updated: 2019/12/08 17:44:16 by chamada     ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/12/17 18:05:59 by chamada     ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -24,8 +24,8 @@ typedef enum		e_cardinal
 
 typedef struct		s_vector
 {
-	double		x;
-	double		y;
+	float		x;
+	float		y;
 }					t_vector;
 
 typedef struct		s_ivector
@@ -34,10 +34,10 @@ typedef struct		s_ivector
 	int	y;
 }					t_ivector;
 
-t_vector			vector(double x, double y) __attribute__((always_inline));
+t_vector			vector(float x, float y) __attribute__((always_inline));
 t_vector			cardinal(t_cardinal direction);
 t_ivector			ivector(int x, int y) __attribute__((always_inline));
 t_ivector			vtoiv(t_vector vector) __attribute__((always_inline));
-void				vrotate(t_vector *vector, double alpha);
+void				vrotate(t_vector *vector, float alpha);
 
 #endif
